+++ packages/features_shared/lib/src/auth/presentation/auth_routes.dart
@@ -1,9 +1,8 @@
 import 'package:go_router/go_router.dart';
 
-import '../../onboarding/presentation/onboarding_screen.dart';
-import '../../onboarding/presentation/splash_screen.dart';
 import 'complete_profile_screen.dart';
 import 'login_screen.dart';
+import 'splash_screen.dart';
 
 const authLoginPath = '/login';
 const completeProfilePath = '/complete-profile';
@@ -12,10 +11,6 @@ final List<GoRoute> authRoutes = [
   GoRoute(
     path: '/splash',
     builder: (context, state) => const SplashScreen(),
-  ),
-  GoRoute(
-    path: '/onboarding',
-    builder: (context, state) => const OnboardingScreen(),
   ),
   GoRoute(
     path: authLoginPath,